--- a/rllib/algorithms/algorithm_config.py
+++ b/rllib/algorithms/algorithm_config.py
@@ -2560,31 +2560,31 @@ class AlgorithmConfig(_Config):
                 `MultiAgentEpisode` not supported, yet). Note,
                 `rllib.core.columns.Columns.OBS` will also try to decompress
                 `rllib.core.columns.Columns.NEXT_OBS`.
-            materialize_data: Whether the raw data should be materialized in memory. This
-                boosts performance, but requires enough memory to avoid an OOM, so
+            materialize_data: Whether the raw data should be materialized in memory.
+                This boosts performance, but requires enough memory to avoid an OOM, so
                 make sure that your cluster has the resources available. For very large
                 data you might want to switch to streaming mode by setting this to
                 `False` (default). If your algorithm does not need the RLModule in the
                 Learner connector pipeline or all (learner) connectors are stateless
                 you should consider setting `materialize_mapped_data` to `True`
-                instead (and set `materialize_data` to `False`).
-                If your data does not fit into memory and your Learner connector pipeline
-                requires an RLModule or is stateful, set both
-                `materialize_data` and `materialize_mapped_data` to `False`.
+                instead (and set `materialize_data` to `False`). If your data does not
+                fit into memory and your Learner connector pipeline requires an RLModule
+                or is stateful, set both `materialize_data` and
+                `materialize_mapped_data` to `False`.
             materialize_mapped_data: Whether the data should be materialized after
-                running it through the Learner connector pipeline (i.e. after running the
-                `OfflinePreLearner`). This improves performance, but should only be used
-                in case the (learner) connector pipeline does not require an RLModule
-                and the (learner) connector pipeline is stateless. For example, MARWIL's
-                Learner connector pipeline requires the RLModule for value function
-                predictions and training batches would become stale after some iterations
-                causing learning degradation or divergence. Also ensure that your cluster
-                has enough memory available to avoid an OOM. If set to `True` (True), make
-                sure that `materialize_data` is set to `False` to avoid materialization of two
-                datasets.
-                If your data does not fit into memory and your Learner connector pipeline
-                requires an RLModule or is stateful, set both
-                `materialize_data` and `materialize_mapped_data` to `False`.
+                running it through the Learner connector pipeline (i.e. after running
+                the `OfflinePreLearner`). This improves performance, but should only be
+                used in case the (learner) connector pipeline does not require an
+                RLModule and the (learner) connector pipeline is stateless. For example,
+                MARWIL's Learner connector pipeline requires the RLModule for value
+                function predictions and training batches would become stale after some
+                iterations causing learning degradation or divergence. Also ensure that
+                your cluster has enough memory available to avoid an OOM. If set to
+                `True` (True), make sure that `materialize_data` is set to `False` to
+                avoid materialization of two datasets. If your data does not fit into
+                memory and your Learner connector pipeline requires an RLModule or is
+                stateful, set both `materialize_data` and `materialize_mapped_data` to
+                `False`.
             map_batches_kwargs: Keyword args for the `map_batches` method. These will be
                 passed into the `ray.data.Dataset.map_batches` method when sampling
                 without checking. If no arguments passed in the default arguments