--- a/rllib/core/learner/learner.py
+++ b/rllib/core/learner/learner.py
@@ -1093,7 +1093,6 @@ class Learner(Checkpointable):
             )
 
         self._check_is_built()
-        minibatch_size = minibatch_size or 1
 
         # Call `before_gradient_based_update` to allow for non-gradient based
         # preparations-, logging-, and update logic to happen.