## Why are these changes needed?

The `map_batches` call in offline RL learning used to be very slow for unknown reasons. This PR proposes multiple changes to the offline data pipeline to boost performance by several multitudes. These changes are

- Materialization of raw data in memory if resources are available with the option `materialize_data` (default is `False`) such that users can control memory usage.
- Materialization of mapped data in memory if resources are available with the option `materialize_mapped_data` (default is `True`) such htat users can control memory usage. This materialization applies the `OfflinePreLearner` on the raw data a priori and can be used by algorithms that do not have connector pipelines (`ConnectorV2` pipelines) that need up-to-date `RLModule` and/or states (e.g. `BC` or `CQL`).
- An iterator that is instantiated once and is reinitiated whenever exhausted for the single-learner case (in the multi-learner case iterators are built on the remote learners anyways)
- A batch size of 1 after the `map_batches` call because rows contain now `MultiAgentBatch`es with `train_batch_size_per_learner` environment steps each.

In addition, it fixes an important error in `MARWIL`'s loss which ignored training the value function.

These changes lead to enormous performance boosts:
- Learning `CartPole-v1` with `BC` in single-learner mode below 7 secs (multi-learner mode < 12 secs).
- Learning `CartPole-v1` with `MARWIL` in single-learner mode below  50 secs (multi-learner mode < 217 secs)
- Learning `Pendulum-v1" with `CQL` in single-learner mode below 311 secs (multi-learner mode < 116 secs)


## Related issue number



## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [x] I've included any doc changes needed for https://docs.ray.io/en/master/.
    - [ ] I've added any new APIs to the API Reference. For example, if I added a 
           method in Tune, I've added it in `doc/source/tune/api/` under the 
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(